Refactor llm_runner to have chat mode and remove llm_app.py
