+++ python/turbine_models/tests/stateless_llama_test.py
@@ -43,28 +43,28 @@ def test_vmfb_comparison():
     quantization = "unquantized"
     precision = "f32"
 
-    llama.export_transformer_model(
-        hf_model_name="Trelis/Llama-2-7b-chat-hf-function-calling-v2",
-        hf_auth_token=None,
-        compile_to="vmfb",
-        external_weights="safetensors",
-        # external_weight_file="Llama-2-7b-chat-hf-function-calling-v2_f16_int4.safetensors", Do not export weights because this doesn't get quantized
-        quantization=quantization,
-        precision=precision,
-        device="llvm-cpu",
-        target_triple="host",
-    )
+    # llama.export_transformer_model(
+    #     hf_model_name="Trelis/Llama-2-7b-chat-hf-function-calling-v2",
+    #     hf_auth_token=None,
+    #     compile_to="vmfb",
+    #     external_weights="safetensors",
+    #     # external_weight_file="Llama-2-7b-chat-hf-function-calling-v2_f16_int4.safetensors", Do not export weights because this doesn't get quantized
+    #     quantization=quantization,
+    #     precision=precision,
+    #     device="llvm-cpu",
+    #     target_triple="host",
+    # )
 
-    from turbine_models.gen_external_params.gen_external_params import (
-        gen_external_params,
-    )
+    # from turbine_models.gen_external_params.gen_external_params import (
+    #     gen_external_params,
+    # )
 
-    gen_external_params(
-        hf_model_name="Trelis/Llama-2-7b-chat-hf-function-calling-v2",
-        quantization=quantization,
-        hf_auth_token=None,
-        precision=precision,
-    )
+    # gen_external_params(
+    #     hf_model_name="Trelis/Llama-2-7b-chat-hf-function-calling-v2",
+    #     quantization=quantization,
+    #     hf_auth_token=None,
+    #     precision=precision,
+    # )
 
     DEFAULT_PROMPT = """<s>[INST] <<SYS>>
 Be concise. You are a helpful, respectful and honest assistant. If a question does not make any sense, or is not factually coherent, explain why instead of answering something not correct. If you don't know the answer to a question, please don't share false information. <</SYS>> hi what are you? [/INST]
